+++ redis-cache/src/main/java/com/example/rediscache/repository/mongo/PopUpStoreMongoRepository.java
@@ -1,8 +1,12 @@
-package com.example.rediscache.repository;
+package com.example.rediscache.repository.mongo;
+
+import java.time.LocalDate;
+import java.util.List;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.example.rediscache.domain.PopUpStoreInfo;
 
 public interface PopUpStoreMongoRepository extends MongoRepository<PopUpStoreInfo, String> {
+	List<PopUpStoreInfo> findByUpdatedDate(LocalDate yourDateField);
 }
